Fixed bug with file export clipped to 255
Typevector isn't converted to uint8 anymore, instead it is single as all the other outputs.
If you concatenate arrays in Matlab that contain different types (uint8 and single), then Matlab decides to convert everything to uint8. Doesn't make sense to me?

+++ piv_FFTmulti.m
@@ -159,7 +159,7 @@ for multipass = 1:passes
 			interrogationarea_center = double((interrogationarea+1)/2);
 		end
 
-		typevector = ones(numelementsy, numelementsx, 'uint8');
+		typevector = ones(numelementsy, numelementsx, 'single');
 		if multipass == 1
 			xtable = zeros(numelementsy,numelementsx, 'single');
 			ytable = zeros(numelementsy,numelementsx, 'single');
